# node-portal/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Backend Apis",
    "description": "API documentation for managing hotel booking requirement"
  },
  "basePath": "/api",
  "schemes": ["http", "https"],
  "paths": {
    "/room/all": {
      "get": {
        "summary": "Get all rooms",
        "parameters": [
          {
            "name": "propertyKey",
            "in": "query",
            "description": "Property key",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "object",
              "properties": {
                "status": {
                  "type": "boolean"
                },
                "data": {
                  "type": "array",
                  "items": {
                    "$ref": "#/definitions/Room"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  "/admin/login": {
    "post": {
       "tags": [
          "Admin"
        ],
      "summary": "Admin login",
      "consumes": [
        "application/json"
      ],
      "parameters": [
        {
          "name": "body",
          "in": "body",
          "description": "Credentials",
          "required": true,
          "schema": {
            "$ref": "#/definitions/AdminLoginRequest"
          }
        }
      ],
      "responses": {
        "200": {
          "description": "OK",
          "schema": {
            "type": "object",
            "properties": {
              "status": {
                "type": "boolean"
              },
              "data": {
                "type": "object",
                "properties": {
                  "user": {
                    "$ref": "#/definitions/User"
                  },
                  "token": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
    "/admin/monthlytaxreport/{date}": {
      "get": {
        "tags":[
          "Admin"
        ],
        "summary": "Get monthly tax report by date",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "The date in the format 'YYYY-MM-DD'",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
         "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "description": "Status of the request"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "date": {
                            "type": "string",
                            "format": "date",
                            "description": "Date of the tax report"
                          },
                          "chargebeforetaxes": {
                            "type": "string",
                            "description": "Charge before taxes"
                          },
                          "kstandard": {
                            "type": "string",
                            "description": "K Standard amount"
                          },
                          "single": {
                            "type": "string",
                            "description": "Single amount"
                          },
                          "queen": {
                            "type": "string",
                            "description": "Queen amount"
                          },
                          "weeklySingle": {
                            "type": "string",
                            "description": "Weekly Single amount"
                          },
                          "standard": {
                            "type": "string",
                            "description": "Standard amount"
                          },
                          "newtaxtype": {
                            "type": "string",
                            "description": "New tax type amount"
                          },
                          "kingsuiteweekly": {
                            "type": "string",
                            "description": "King Suite Weekly amount"
                          },
                          "doubledaily": {
                            "type": "string",
                            "description": "Double Daily amount"
                          },
                          "doubleweekly": {
                            "type": "string",
                            "description": "Double Weekly amount"
                          },
                          "singleQueenDaily": {
                            "type": "string",
                            "description": "Single Queen Daily amount"
                          },
                          "singleQueenWeekly": {
                            "type": "string",
                            "description": "Single Queen Weekly amount"
                          },
                          "taxTotal": {
                            "type": "string",
                            "description": "Total tax amount"
                          },
                          "taxExemptTotal": {
                            "type": "string",
                            "description": "Total tax exempt amount"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
     "/admin/monthlysalesreport/{date}": {
      "get": {
        "tags":[
          "Admin"
        ],
        "summary": "Get monthly sales report by date",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "The date in the format 'YYYY-MM-DD'",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": {
                      "type": "boolean",
                      "description": "Status of the request"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": {
                            "type": "integer",
                            "description": "Name"
                          },
                          "description": {
                            "type": "string",
                            "description": "Description"
                          },
                          "_": {
                            "type": "string",
                            "description": "_"
                          },
                          "percentage": {
                            "type": "string",
                            "description": "Percentage"
                          },
                          "nights": {
                            "type": "string",
                            "description": "Nights"
                          },
                          "room": {
                            "type": "string",
                            "description": "Room"
                          },
                          "other": {
                            "type": "string",
                            "description": "Other"
                          },
                          "tax": {
                            "type": "string",
                            "description": "Tax"
                          },
                          "total": {
                            "type": "string",
                            "description": "Total"
                          },
                          "payment": {
                            "type": "string",
                            "description": "Payment"
                          },
                          "balance": {
                            "type": "string",
                            "description": "Balance"
                          },
                          "cancel": {
                            "type": "string",
                            "description": "Cancel"
                          },
                          "avgstay": {
                            "type": "string",
                            "description": "Average Stay"
                          },
                          "adr": {
                            "type": "string",
                            "description": "ADR"
                          },
                          "revpar": {
                            "type": "string",
                            "description": "RevPAR"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
           "401": {
            "description": "Access denied. No token provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Access denied. No token provided."
                }
              }
            }
          }
        }
      }
    },
    "/admin/summaryreport/{date}": {
      "get": {
        "tags":[
          "Admin"
        ],
        "summary": "Get summary report",
        "description": "Retrieves the summary report for the specified date.",
        "parameters": [
          {
            "name": "date",
            "in": "path",
            "description": "The date of the summary report (e.g., '2023-05-25')",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryReportResponse"
                }
              }
            }
          },
          "401": {
            "description": "Access denied. No token provided.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "message": "Access denied. No token provided."
                }
              }
            }
          }
        }
      }
    },
...
  },
...
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
 "securityDefinitions": {
    "BearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "definitions": {
    "Room": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "roomName": {
          "type": "string"
        },
        "minimumNightStays": {
          "type": "integer"
        },
        "maximumOccupancy": {
          "type": "integer"
        },
        "pricePerNight": {
          "type": "integer"
        },
        "totalNumberOfRooms": {
          "type": "integer"
        },
        "description": {
          "type": "string"
        },
        "amenities": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "images": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "imageKey": {
                "type": "string"
              },
              "imagePath": {
                "type": "string"
              }
            }
          }
        },
        "totalNumberOfBookedRooms": {
          "type": "integer"
        },
        "totalNOfRooms": {
          "type": "integer"
        },
        "iconName": {
          "type": "string"
        },
        "propertyKey": {
          "type": "string"
        },
        "roomType": {
          "type": "string"
        }
      }
    },
...
  "AdminLoginRequest": {
    "type": "object",
    "properties": {
      "email": {
        "type": "string"
      },
      "password": {
        "type": "string"
      }
    }
  },
...
  }
}